[Feat][Core/Dashboard] Merge APIHead into JobHead and ReportHead (#52389)
