--- python/ray/dashboard/modules/api/tests/conftest.py
@@ -1,1 +1,0 @@
-from ray._private.conftest_utils import set_override_dashboard_url  # noqa: F401
