+++ python/ray/dashboard/modules/job/component_activities_schema.json
@@ -1,6 +1,6 @@
 {
   "$schema": "http://json-schema.org/draft-07/schema#",
-  "$id": "http://github.com/ray-project/ray/dashboard/modules/snapshot/component_activities_schema.json",
+  "$id": "http://github.com/ray-project/ray/dashboard/modules/job/component_activities_schema.json",
   "type": "object",
   "patternProperties": {
     "[0-9a-f]*": {
